requirements.txt
Django==2.2.1
django-model-utils==3.1.2
djangorestframework==3.8.2
pytz==2019.1
sqlparse==0.3.0
psycopg2-binary==2.8.4
django-filter==21.1

# debug and testing
django-debug-toolbar==3.2.4
coverage==7.0.0
pytest==7.2.0
pytest-django==4.5.2
factory-boy==3.2.1

# linters and formatters
black>=22.3.0
flake8>=4.0.1
autoflake>=2.0.0
mypy>=0.942
isort>=5.10.1
